link and column updates
