+++ apps/platform/src/pages/StudyPage/Header.tsx
@@ -4,41 +4,49 @@ import { Header as HeaderBase, ExternalLink, XRefLinks } from "ui";
 type HeaderProps = {
   loading: boolean;
   studyId: string;
-  traitFromSource: string;
-  backgroundTraits: any[];  // array of diseases - wait until get types directly from schema
-  targetId: any;  // target - wait until get types directly from schema
-  diseaseId: string;
+  backgroundTraits: {
+    id: string;
+    name: string;
+  }[];
+  targetId: string;
+  diseases: {
+    id: string;
+    name: string;
+  }[];
   studyCategory: string;
 };
 
-function Header({ loading,
-                  studyId,
-                  traitFromSource,
-                  backgroundTraits,
-                  targetId,
-                  diseaseId,
-                  studyCategory }: HeaderProps) {
+function Header({
+      loading,
+      studyId,
+      backgroundTraits,
+      targetId,
+      diseases,
+      studyCategory
+    }: HeaderProps) {
 
-  let traitLink, sourceLink;
+  let traitLinks, sourceLink;
   if (studyCategory === 'GWAS') {
-    if (diseaseId) {
-      traitLink = {
-        title: "Trait",
-        id: diseaseId,
-        url: `../disease/${diseaseId}`, 
-      };
+    if (diseases?.length) {
+      traitLinks = <XRefLinks
+        label="Trait"
+        urlStem="../disease/"
+        ids={diseases.map(d => d.id)}
+        names={diseases.map(d => d.name)}
+      />
     }
     sourceLink = {
       id: "GWAS Catalog",
       url: `https://www.ebi.ac.uk/gwas/studies/${studyId}`,
     };
   } else if (studyCategory === 'FINNGEN') {
-    if (diseaseId) {
-      traitLink = {
-        title: "Trait",
-        id: diseaseId,
-        url: `../disease/${diseaseId}`,
-      };
+    if (diseases?.length) {
+      traitLinks = <XRefLinks
+        label="Trait"
+        urlStem="../disease/"
+        ids={diseases.map(d => d.id)}
+        names={diseases.map(d => d.name)}
+      />
     }
     sourceLink = {
       id: "FinnGenR10",
@@ -46,11 +54,11 @@ function Header({ loading,
     };
   } else {  // QTL
     if (targetId) {
-      traitLink = {
-        title: "Affected gene",
-        id: targetId,
-        url: `../target/${targetId}`,
-      };
+      traitLinks = <ExternalLink
+        title="Affected gene"
+        id={targetId}
+        url={`../target/${targetId}`}
+      />
     }
     sourceLink = {
       id: "eQTL Catalog",
@@ -65,28 +73,23 @@ function Header({ loading,
       subtitle={null}
       Icon={faChartBar}
       externalLinks={
-        traitLink &&
-          <> 
-            <ExternalLink
-              title={traitLink.title}
-              id={traitFromSource ?? traitLink.id}
-              url={traitLink.url}
-            />
-            { 
-              studyCategory === "GWAS" && backgroundTraits?.length > 0 &&
-                <XRefLinks
-                  label="Background traits"
-                  urlStem="../disease/"
-                  ids={backgroundTraits.map(t => t.id)}
-                  names={backgroundTraits.map(t => t.name)}
-                />
-            }
-            <ExternalLink
-              title="Source"
-              id={sourceLink.id}
-              url={sourceLink.url}
-            />
-          </>
+        <> 
+          {traitLinks}
+          { 
+            studyCategory === "GWAS" && backgroundTraits?.length > 0 &&
+              <XRefLinks
+                label="Background traits"
+                urlStem="../disease/"
+                ids={backgroundTraits.map(t => t.id)}
+                names={backgroundTraits.map(t => t.name)}
+              />
+          }
+          <ExternalLink
+            title="Source"
+            id={sourceLink.id}
+            url={sourceLink.url}
+          />
+        </>
       }
     />
   );
