+++ apps/platform/src/pages/StudyPage/Profile.tsx
@@ -8,20 +8,20 @@ import {
   summaryUtils,
 } from "ui";
 
-import RelatedGWASStudiesSummary from "sections/src/study/RelatedGWASStudies/Summary";
+import SharedTraitStudiesSummary from "sections/src/study/SharedTraitStudies/Summary";
 import GWASCredidbleSetsSummary from "sections/src/study/GWASCredibleSets/Summary";
 import QTLCredibleSetsSummary from "sections/src/study/QTLCredibleSets/Summary";
 
 import client from "../../client";
 import ProfileHeader from "./ProfileHeader";
 
-const RelatedGWASStudiesSection = lazy(
-  () => import("sections/src/study/RelatedGWASStudies/Body")
+const SharedTraitStudiesSection = lazy(
+  () => import("sections/src/study/SharedTraitStudies/Body")
 );
 const GWASCredibleSetsSection = lazy(() => import("sections/src/study/GWASCredibleSets/Body"));
 const QTLCredibleSetsSection = lazy(() => import("sections/src/study/QTLCredibleSets/Body"));
 
-// no RelatedGWASStudiesSummary as we add section to the query below directly
+// no SharedTraitStudiesSummary as we add section to the query below directly
 // (the summary cannot be written as a fragment as it gets further studies)
 const summaries = [GWASCredidbleSetsSummary, QTLCredibleSetsSummary];
 
@@ -32,13 +32,13 @@ const STUDY_PROFILE_SUMMARY_FRAGMENT = summaryUtils.createSummaryFragment(
   "StudyProfileSummaryFragment"
 );
 const STUDY_PROFILE_QUERY = gql`
-  query StudyProfileQuery($studyId: String!, $diseaseId: String!) {
+  query StudyProfileQuery($studyId: String!, $diseaseIds: [String!]!) {
     gwasStudy(studyId: $studyId) {
       studyId
       ...StudyProfileHeaderFragment
       ...StudyProfileSummaryFragment
     }
-    relatedGWASStudies: gwasStudy(diseaseId: $diseaseId, page: { size: 2, index: 0}) {
+    sharedTraitStudies: gwasStudy(diseaseIds: $diseaseIds, page: { size: 2, index: 0}) {
       studyId
     }
   }
@@ -49,11 +49,14 @@ const STUDY_PROFILE_QUERY = gql`
 type ProfileProps = {
   studyId: string;
   studyCategory: string;
-  diseaseIds: string[];
+  diseases: {
+    id: string;
+    name: string;
+  }[];
 };
 
-function Profile({ studyId, studyCategory, diseaseIds }: ProfileProps) {
-  const diseaseId = diseaseIds?.[0] || "";  // !! WILL LEAVE AS diseaseIds WHEN API UPDATED !!
+function Profile({ studyId, studyCategory, diseases }: ProfileProps) {
+  const diseaseIds = diseases?.map(d => d.id) || [];
 
   return (
     <PlatformApiProvider
@@ -61,7 +64,7 @@ function Profile({ studyId, studyCategory, diseaseIds }: ProfileProps) {
       query={STUDY_PROFILE_QUERY}
       variables={{
         studyId,
-        diseaseId,  // !! WILL BE diseaseIds WHEN API UPDATED !!
+        diseaseIds,
       }}
       client={client}
     >
@@ -70,7 +73,7 @@ function Profile({ studyId, studyCategory, diseaseIds }: ProfileProps) {
       <SummaryContainer>
         {(studyCategory === "GWAS" || studyCategory === "FINNGEN") &&
           <>
-            <RelatedGWASStudiesSummary />
+            <SharedTraitStudiesSummary />
             <GWASCredidbleSetsSummary />
           </>
         }
@@ -83,9 +86,9 @@ function Profile({ studyId, studyCategory, diseaseIds }: ProfileProps) {
         {(studyCategory === "GWAS" || studyCategory === "FINNGEN") &&
           <>
             <Suspense fallback={<SectionLoader />}>
-              <RelatedGWASStudiesSection
+              <SharedTraitStudiesSection
                 studyId={studyId}
-                diseaseId={diseaseId}  // !! WILL BE diseaseIds WHEN API UPDATED !!
+                diseaseIds={diseaseIds}
                 entity={STUDY}
               />
             </Suspense>
