+++ apps/platform/src/pages/StudyPage/ProfileHeader.gql
@@ -4,8 +4,6 @@ fragment StudyProfileHeaderFragment on Gwas {
   publicationDate
   publicationJournal
   pubmedId
-  hasSumstats
-  summarystatsLocation
   nSamples
   initialSampleSize
   replicationSamples {
