+++ apps/platform/src/pages/StudyPage/StudyPage.gql
@@ -2,7 +2,6 @@ query StudyPageQuery($studyId: String!) {
   gwasStudy(studyId: $studyId) {
     studyId
     projectId
-    traitFromSource
     backgroundTraits {
       id
       name
@@ -12,6 +11,7 @@ query StudyPageQuery($studyId: String!) {
     }
     diseases {
       id
+      name
     }
   }
 }
