+++ apps/platform/src/pages/StudyPage/index.ts
