--- packages/sections/src/study/RelatedGWASStudies/RelatedGWASStudiesQuery.gql
@@ -1,28 +1,0 @@
-query RelatedGWASStudiesQuery($diseaseId: String!) {
-  gwasStudy(diseaseId: $diseaseId, page: { size: 2000, index: 0 }) {
-    studyId
-    projectId
-    publicationFirstAuthor
-    publicationDate
-    publicationJournal
-    nSamples
-    initialSampleSize
-    discoverySamples {
-      ancestry
-      sampleSize
-    }
-    replicationSamples {
-      ancestry
-      sampleSize
-    }
-    nCases
-    nControls
-    cohorts
-    ldPopulationStructure {
-      ldPopulation
-      relativeSampleSize
-    }
-    hasSumstats
-    pubmedId
-  }
-}
