+++ packages/sections/src/study/SharedTraitStudies/Description.tsx
@@ -7,14 +7,14 @@ type DescriptionProps = {
 function Description({ studyId }: DescriptionProps) {
   return (
     <>
-      GWAS studies that share associated traits with study{" "}
+      GWAS studies that share traits with study{" "}
       <strong>{studyId}</strong>. Source{" "}
       <Link external to="https://www.ebi.ac.uk/gwas/studies" >
         GWAS Catalog
       </Link>
       ,{" "}
       <Link external to="https://r10.finngen.fi" >
-        FinnGenR10
+        FinnGen
       </Link>
       .
     </>
