+++ packages/sections/src/study/SharedTraitStudies/SharedTraitStudiesQuery.gql
@@ -1,0 +1,20 @@
+query SharedTraitStudiesQuery($diseaseIds: [String!]!) {
+  gwasStudy(diseaseIds: $diseaseIds, page: { size: 2000, index: 0 }) {
+    studyId
+    projectId
+    diseases {
+      id
+      name
+    }
+    publicationFirstAuthor
+    publicationDate
+    publicationJournal
+    nSamples
+    cohorts
+    ldPopulationStructure {
+      ldPopulation
+      relativeSampleSize
+    }
+    pubmedId
+  }
+}
