+++ packages/sections/src/study/SharedTraitStudies/Summary.tsx
